+++ TMDB-SwiftUI/Services/Database/Models/ProductionCompany+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension ProductionCompany {
         return NSFetchRequest<ProductionCompany>(entityName: "ProductionCompany")
     }
 
-    @NSManaged public var id: Int16
+    @NSManaged public var id: Int64
     @NSManaged public var logoPath: String?
     @NSManaged public var name: String?
     @NSManaged public var movie: MovieDetailsResponse?
@@ -24,5 +24,4 @@ extension ProductionCompany {
 }
 
 extension ProductionCompany : Identifiable {
-
 }
